Move routes definitions to separate router.
